requirements.txt
python-daemon
pycryptodome